+++ modules/user/classes/PreferenceMgr.php
@@ -1,7 +1,7 @@
 <?php
 /* Reminder: always indent with 4 spaces (no tabs). */
 // +---------------------------------------------------------------------------+
-// | Copyright (c) 2005, Demian Turner                                         |
+// | Copyright (c) 2006, Demian Turner                                         |
 // | All rights reserved.                                                      |
 // |                                                                           |
 // | Redistribution and use in source and binary forms, with or without        |
@@ -30,7 +30,7 @@
 // | OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.      |
 // |                                                                           |
 // +---------------------------------------------------------------------------+
-// | Seagull 0.4                                                               |
+// | Seagull 0.6                                                               |
 // +---------------------------------------------------------------------------+
 // | PreferenceMgr.php                                                         |
 // +---------------------------------------------------------------------------+
@@ -39,15 +39,14 @@
 // $Id: PreferenceMgr.php,v 1.39 2005/05/17 23:54:53 demian Exp $
 
 require_once SGL_MOD_DIR . '/user/classes/DA_User.php';
+require_once 'DB/DataObject.php';
 
 /**
  * Manages user permissions.
  *
  * @package User
  * @author  Demian Turner <demian@phpkitchen.com>
- * @copyright Demian Turner 2004
  * @version $Revision: 1.39 $
- * @since   PHP 4.1
  */
 class PreferenceMgr extends SGL_Manager
 {
@@ -55,18 +54,18 @@ class PreferenceMgr extends SGL_Manager
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
         parent::SGL_Manager();
-        $this->module       = 'user';
+
         $this->template     = 'prefManager.html';
         $this->pageTitle    = 'Preference Manager';
         $this->da           = & DA_User::singleton();
 
         $this->_aActionsMapping =  array(
-            'add'       => array('add'), 
+            'add'       => array('add'),
             'insert'    => array('insert', 'redirectToDefault'),
-            'edit'      => array('edit'), 
-            'update'    => array('update', 'redirectToDefault'), 
-            'delete'    => array('delete', 'redirectToDefault'), 
-            'list'      => array('list'), 
+            'edit'      => array('edit'),
+            'update'    => array('update', 'redirectToDefault'),
+            'delete'    => array('delete', 'redirectToDefault'),
+            'list'      => array('list'),
         );
 
         $this->aThemes = SGL_Util::getAllThemes();
@@ -75,6 +74,7 @@ class PreferenceMgr extends SGL_Manager
             'US' => 'US',
             'FR' => 'FR',
             'BR' => 'BR',
+            'DE' => 'DE',
             );
         $this->aTimeouts = array(
             '900' => '15 mins',
@@ -95,11 +95,12 @@ class PreferenceMgr extends SGL_Manager
     function validate($req, &$input)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
+
         $this->validated        = true;
         $input->pageTitle       = $this->pageTitle;
         $input->masterTemplate  = 'masterMinimal.html';
         $input->template        = $this->template;
-        $input->submit          = $req->get('submitted');
+        $input->submitted       = $req->get('submitted');
         $input->action          = ($req->get('action')) ? $req->get('action') : 'list';
         $input->from            = ($req->get('frmFrom'))?$req->get('frmFrom'):0;
         $input->prefId          = $req->get('frmPrefId');
@@ -107,13 +108,18 @@ class PreferenceMgr extends SGL_Manager
         $input->pref            = (object) $req->get('pref');
         $input->aDelete         = $req->get('frmDelete');
         $input->totalItems      = $req->get('totalItems');
+        $input->sortBy          = SGL_Util::getSortBy($req->get('frmSortBy'), SGL_SORTBY_USER);
+        $input->sortOrder       = SGL_Util::getSortOrder($req->get('frmSortOrder'));
+
+        // This will tell HTML_Flexy which key is used to sort data
+        $input->{ 'sort_' . $input->sortBy } = true;
 
         $aErrors = array();
-        if ($input->submit) {
+        if ($input->submitted || in_array($input->action, array('insert', 'update'))) {
             if (empty($input->pref->name)) {
                 $aErrors['name'] = 'You must enter a preference name';
             }
-            if (is_null($input->pref->default_value) || !strlen($input->pref->default_value)) {
+            if (empty($input->pref->default_value)) {
                 $aErrors['default_value'] = 'You must enter a default value';
             }
         }
@@ -121,90 +127,95 @@ class PreferenceMgr extends SGL_Manager
         if (is_array($aErrors) && count($aErrors)) {
             SGL::raiseMsg('Please fill in the indicated fields');
             $input->error = $aErrors;
+
             if ($input->action == 'insert') {
-                $input->pageTitle = $this->pageTitle . ' :: Add';
-                $input->template = 'prefAdd.html';
+                $input->preferenceAdd = true;
             } else {
-                $input->pageTitle = $this->pageTitle . ' :: Edit';
-                $input->template = 'prefEdit.html';
+                $input->preferenceEdit = true;
             }
+            $input->template = 'prefEdit.html';
             $this->validated = false;
         }
     }
-    
+
     function display(&$output)
     {
-        //  load available languages
-        $availableLanguages = $GLOBALS['_SGL']['LANGUAGE'];
-        uasort($availableLanguages, 'SGL_cmp');
-        foreach ($availableLanguages as $id => $tmplang) {
-            $lang_name = ucfirst(substr(strstr($tmplang[0], '|'), 1));
-            $aLangOptions[$id] =  $lang_name . ' (' . $id . ')';
-        }
-        $output->aLangs = $aLangOptions;
-        
+        //  build map of available languages
+        $output->aLangs = SGL_Util::getLangsDescriptionMap();
+
         //  FIXME: unix-only, create fallback for windows
         $locales = explode("\n", @shell_exec('locale -a'));
 
         //  remap to hash
         foreach ($locales as $locale) {
             $aLocales[$locale] = $locale;
-        }        
+        }
         $output->aLocales = $aLocales;
         require_once SGL_DAT_DIR . '/ary.timezones.en.php';
-        $output->aTimezones = $tz;        
+        $output->aTimezones = $tz;
     }
 
-    function _add(&$input, &$output)
+    function _cmd_add(&$input, &$output)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
-        require_once SGL_ENT_DIR . '/Preference.php';
-        $output->template = 'prefAdd.html';
+
+        $output->template = 'prefEdit.html';
+        $output->preferenceAdd = true;
         $output->pageTitle = $this->pageTitle . ' :: Add';
-        $output->pref = & new DataObjects_Preference();
+        $output->pref = DB_DataObject::factory($this->conf['table']['preference']);
     }
 
-    function _insert(&$input, &$output)
+    function _cmd_insert(&$input, &$output)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
-        require_once SGL_ENT_DIR . '/Preference.php';
-        $oPref = & new DataObjects_Preference();
+
+        SGL_DB::setConnection();
+        $oPref = DB_DataObject::factory($this->conf['table']['preference']);
         $oPref->setFrom($input->pref);
-        $dbh = & $oPref->getDatabaseConnection();
-        $conf = & $GLOBALS['_SGL']['CONF'];
-        
-        $oPref->preference_id = $dbh->nextId($conf['table']['preference']);
+
+        $oPref->preference_id = $this->dbh->nextId($this->conf['table']['preference']);
         $success = $oPref->insert();
         if ($success) {
-            //  synchronise with user_preference table
-            $ret = $this->da->syncDefaultPrefs();
-            SGL::raiseMsg('pref successfully added');
+            // add new preference to all users prefs
+            $oUser = DB_DataObject::factory($this->conf['table']['user']);
+            $oUser->find();
+            while ($oUser->fetch()) {
+                $ret = $this->da->addPrefsByUserId(
+                    array($oPref->preference_id => $oPref->default_value), $oUser->usr_id);
+            }
+            SGL::raiseMsg('pref successfully added', true, SGL_MESSAGE_INFO);
         } else {
-           SGL::raiseError('There was a problem inserting the record', 
+           SGL::raiseError('There was a problem inserting the record',
                 SGL_ERROR_NOAFFECTEDROWS);
         }
     }
 
-    function _edit(&$input, &$output)
+    function _cmd_edit(&$input, &$output)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
-        require_once SGL_ENT_DIR . '/Preference.php';
-        $output->template = 'prefEdit.html';
-        $output->pageTitle = $this->pageTitle . ' :: Edit';
-        $oPref = & new DataObjects_Preference();
-        $oPref->get($input->prefId);
-        $output->pref = $oPref;
+
+        if (is_null($input->prefId)) {
+            SGL::raiseError('Incorrect parameter passed to ' . __CLASS__ . '::' .
+                __FUNCTION__, SGL_ERROR_INVALIDARGS);
+        } else {
+            $output->preferenceEdit = true;
+            $output->template = 'prefEdit.html';
+            $output->pageTitle = $this->pageTitle . ' :: Edit';
+            $oPref = DB_DataObject::factory($this->conf['table']['preference']);
+            $oPref->get($input->prefId);
+            $output->pref = $oPref;
+        }
     }
 
-    function _update(&$input, &$output)
+    function _cmd_update(&$input, &$output)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
-        require_once SGL_ENT_DIR . '/Preference.php';
-        $oPref = & new DataObjects_Preference();
+
+        $oPref = DB_DataObject::factory($this->conf['table']['preference']);
         $oPref->get($input->pref->preference_id);
         $oPref->setFrom($input->pref);
         unset($oPref->name);
-        
+
         //  don't check for success because pref.name must remain the same
         $changed = $oPref->update();
 
@@ -212,55 +223,75 @@ class PreferenceMgr extends SGL_Manager
         if ($changed) {
             $ret = $this->da->syncDefaultPrefs();
         }
-        SGL::raiseMsg('pref successfully updated');
+        SGL::raiseMsg('pref successfully updated', true, SGL_MESSAGE_INFO);
     }
 
-    function _delete(&$input, &$output)
+    function _cmd_delete(&$input, &$output)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
-        require_once SGL_ENT_DIR . '/Preference.php';
-        require_once SGL_ENT_DIR . '/User_preference.php';
+
         $aToDelete = array();
-        foreach ($input->aDelete as $index => $prefId) {
-            $oPref = & new DataObjects_Preference();
-            $oPref->get($prefId);
-            $oPref->delete();
-            $aToDelete[] = $prefId;
-            unset($oPref);
+        if (is_array($input->aDelete)) {
+            foreach ($input->aDelete as $index => $prefId) {
+                $oPref = DB_DataObject::factory($this->conf['table']['preference']);
+                $oPref->get($prefId);
+                $oPref->delete();
+                $aToDelete[] = $prefId;
+                unset($oPref);
+            }
+            //  delete related user_prefs
+            foreach ($aToDelete as $deleteId) {
+                $oUserPref = DB_DataObject::factory($this->conf['table']['user_preference']);
+                $oUserPref->get('preference_id', $deleteId);
+                $oUserPref->delete();
+                while ($oUserPref->fetch()) {
+                    $oUserPref->delete();
+                }
+                unset($oUserPref);
+            }
+            SGL::raiseMsg('pref successfully deleted', true, SGL_MESSAGE_INFO);
+        } else {
+            SGL::raiseError('Incorrect parameter passed to ' . __CLASS__ . '::' .
+                __FUNCTION__, SGL_ERROR_INVALIDARGS);
         }
-        //  delete related user_prefs
-        foreach ($aToDelete as $deleteId) {
-            $oUserPref = & new DataObjects_User_preference();
-            $oUserPref->get('preference_id', $deleteId);
-            $oUserPref->delete();
-            unset($oUserPref);
-        }
-        SGL::raiseMsg('pref successfully deleted');
     }
 
-    function _list(&$input, &$output)
+    function _cmd_list(&$input, &$output)
     {
         SGL::logMessage(null, PEAR_LOG_DEBUG);
 
         $output->pageTitle = $this->pageTitle . ' :: Browse';
-        $dbh = & SGL_DB::singleton();
-        $conf = & $GLOBALS['_SGL']['CONF'];
-        
-        $query = "SELECT preference_id, name, default_value FROM {$conf['table']['preference']}";
+
+        $allowedSortFields = array('preference_id','name');
+        if (  !empty($input->sortBy)
+           && !empty($input->sortOrder)
+           && in_array($input->sortBy, $allowedSortFields)) {
+                $orderBy_query = 'ORDER BY ' . $input->sortBy . ' ' . $input->sortOrder ;
+        } else {
+            $orderBy_query = 'ORDER BY preference_id ASC ';
+        }
+
+        $query = "
+            SELECT  preference_id, name, default_value
+            FROM    {$this->conf['table']['preference']}
+            $orderBy_query";
+
         $limit = $_SESSION['aPrefs']['resPerPage'];
         $pagerOptions = array(
-            'mode'      => 'Sliding',
-            'delta'     => 3,
-            'perPage'   => $limit,
-            'totalItems'=> $input->totalItems,
+            'mode'     => 'Sliding',
+            'delta'    => 3,
+            'perPage'  => $limit,
+            'spacesBeforeSeparator' => 0,
+            'spacesAfterSeparator'  => 0,
+            'curPageSpanPre'        => '<span class="currentPage">',
+            'curPageSpanPost'       => '</span>',
         );
-
-        $aPagedData = SGL_DB::getPagedData($dbh, $query, $pagerOptions);
+        $aPagedData = SGL_DB::getPagedData($this->dbh, $query, $pagerOptions);
         $output->aPagedData = $aPagedData;
         if (is_array($aPagedData['data']) && count($aPagedData['data'])) {
             $output->pager = ($aPagedData['totalItems'] <= $limit) ? false : true;
         }
-        $output->addOnLoadEvent("document.getElementById('frmUserMgrChooser').prefs.disabled = true");
+        $output->addOnLoadEvent("switchRowColorOnHover()");
     }
 }
 ?>
